requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.5.30
numpy==1.21.2
pandas==1.3.2
python-dateutil==2.8.2
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2021.1
six==1.16.0
tornado==6.1
tzlocal==3.0
